More qms tests fixes Signed-off-by: Alexey Zinoviev <alexey.zinoviev@xored.com>
