+++ qms-tests/sanity/storageSecond-dev.json
@@ -1,0 +1,34 @@
+{
+  "cookies": [],
+  "origins": [
+    {
+      "origin": "http://localhost:8080",
+      "localStorage": [
+        {
+          "name": "login:metadata:LoginEmail",
+          "value": "user2"
+        },
+        {
+          "name": "login:metadata:LoginTokens",
+          "value": "{\"sanity-ws-qms\":\"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJjb25maXJtZWQiOnRydWUsImVtYWlsIjoidXNlcjIiLCJ3b3Jrc3BhY2UiOiJzYW5pdHktd3MtcW1zIiwicHJvZHVjdElkIjoiIn0.-al3BXM41qvOaZ4s7c96xQZO73udvQrs8dH2bmhScww\"}"
+        },
+        {
+          "name": "login:metadata:LoginEndpoint",
+          "value": "ws://localhost:3334"
+        },
+        {
+          "name": "#platform.notification.logging",
+          "value": "false"
+        },
+        {
+          "name": "#platform.lazy.loading",
+          "value": "false"
+        },
+        {
+          "name": "flagOpenInDesktopApp",
+          "value": "true"
+        }
+      ]
+    }
+  ]
+}
