+++ qms-tests/sanity/tests/documents/documents.spec.ts
@@ -19,6 +19,7 @@ import { DocumentApprovalsPage } from '../model/documents/document-approvals-pag
 import { DocumentReleasePage } from '../model/documents/document-release-page'
 import { DocumentReasonAndImpactPage } from '../model/documents/document-reason-impact-page'
 import { LeftSideMenuPage } from '../model/left-side-menu-page'
+import { DocumentHistoryPage } from '../model/documents/document-history-page'
 
 test.use({
   storageState: PlatformSetting
@@ -330,6 +331,56 @@ test.describe('QMS. Documents tests', () => {
     })
   })
 
+  test('TESTS-137. Approve document', async ({ page }) => {
+    await allure.description('Requirement\nUsers need to approve the document')
+    await allure.tms('TESTS-137', 'https://front.hc.engineering/workbench/platform/tracker/TESTS-137')
+    const approveDocument: NewDocument = {
+      template: 'HR (HR)',
+      title: `Approve document-${generateId()}`,
+      description: `Approve document description-${generateId()}`
+    }
+    const documentDetails: DocumentDetails = {
+      type: 'HR',
+      category: 'Human Resources',
+      version: 'v0.1',
+      status: DocumentStatus.DRAFT,
+      owner: 'Appleseed John',
+      author: 'Appleseed John'
+    }
+
+    await prepareDocumentStep(page, approveDocument)
+    const documentContentPage = new DocumentContentPage(page)
+    await test.step('2. Send for Approval', async () => {
+      await documentContentPage.buttonSendForApproval.click()
+      await documentContentPage.fillSelectApproversForm([documentDetails.owner])
+      await documentContentPage.checkDocumentStatus(DocumentStatus.IN_APPROVAL)
+      await documentContentPage.checkDocument({
+        ...documentDetails,
+        status: DocumentStatus.IN_APPROVAL
+      })
+      await documentContentPage.checkCurrentRights(DocumentRights.VIEWING)
+    })
+
+    await test.step('3. Approve document', async () => {
+      await documentContentPage.confirmApproval()
+    })
+
+    await test.step('4. Check the document and status', async () => {
+      await documentContentPage.checkDocumentStatus(DocumentStatus.EFFECTIVE)
+      await documentContentPage.checkDocument({
+        ...documentDetails,
+        status: DocumentStatus.EFFECTIVE,
+        version: 'v0.1'
+      })
+      await documentContentPage.checkCurrentRights(DocumentRights.VIEWING)
+
+      await documentContentPage.openApprovals()
+      const documentApprovalsPage = new DocumentApprovalsPage(page)
+      await documentApprovalsPage.checkSuccessApproval(documentDetails.owner)
+      await attachScreenshot('TESTS-137_approve_document.png', page)
+    })
+  })
+
   test('TESTS-138. Reject document', async ({ page }) => {
     await allure.description('Requirement\nUsers need to reject the document')
     await allure.tms('TESTS-138', 'https://front.hc.engineering/workbench/platform/tracker/TESTS-138')
@@ -379,13 +430,13 @@ test.describe('QMS. Documents tests', () => {
     })
   })
 
-  test('TESTS-137. Approve document', async ({ page }) => {
-    await allure.description('Requirement\nUsers need to approve the document')
-    await allure.tms('TESTS-137', 'https://front.hc.engineering/workbench/platform/tracker/TESTS-137')
-    const approveDocument: NewDocument = {
+  test('TESTS-139. Make Review', async ({ page }) => {
+    await allure.description('Requirement\nUsers need to make a review for the document')
+    await allure.tms('TESTS-139', 'https://front.hc.engineering/workbench/platform/tracker/TESTS-139')
+    const makeReviewDocument: NewDocument = {
       template: 'HR (HR)',
-      title: `Approve document-${generateId()}`,
-      description: `Approve document description-${generateId()}`
+      title: `Make Review document-${generateId()}`,
+      description: `Make Review document description-${generateId()}`
     }
     const documentDetails: DocumentDetails = {
       type: 'HR',
@@ -395,25 +446,352 @@ test.describe('QMS. Documents tests', () => {
       owner: 'Appleseed John',
       author: 'Appleseed John'
     }
+    const newContentFirst: Content = {
+      sectionTitle: `Make Review. Overview-${generateId()}`,
+      content: `Make Review. New content-${generateId()}!!!!`
+    }
+    const updateContentFirst: Content = {
+      sectionTitle: `Make Review Updated. Updated Overview-${generateId()}`,
+      content: `Make Review Updated. Updated content-${generateId()}!!!!`
+    }
+    const newContentSecond: Content = {
+      sectionTitle: `Make Review. Description-${generateId()}`,
+      content: `Make Review. New content Description-${generateId()}!!!!`
+    }
+    const messageToTitle: string = `Make Review. Message to the first title-${generateId()}`
+    const messageToSecondTitle: string = `Make Review. Message to the second title-${generateId()}`
+    const messageToContent: string = `Make Review. Message to the content-${generateId()}`
 
-    await prepareDocumentStep(page, approveDocument)
+    await prepareDocumentStep(page, makeReviewDocument)
+
     const documentContentPage = new DocumentContentPage(page)
-    await test.step('2. Send for Approval', async () => {
+    await test.step('2. Add section and content', async () => {
+      await documentContentPage.checkDocumentTitle(makeReviewDocument.title)
+      await documentContentPage.checkDocumentStatus(DocumentStatus.DRAFT)
+      await documentContentPage.checkDocument(documentDetails)
+      await documentContentPage.checkCurrentRights(DocumentRights.EDITING)
+
+      await documentContentPage.updateSectionTitle('1', newContentFirst.sectionTitle)
+      await documentContentPage.addContentToTheSection(newContentFirst)
+
+      await documentContentPage.addNewSection('1', 'below')
+      await documentContentPage.updateSectionTitle('2', newContentSecond.sectionTitle)
+      await documentContentPage.addContentToTheSection(newContentSecond)
+      await documentContentPage.checkContentForTheSection(newContentSecond)
+      await attachScreenshot('TESTS-139_add_section_and_content.png', page)
+    })
+
+    await test.step('3. Send for Review', async () => {
+      await documentContentPage.buttonSendForReview.click()
+      await documentContentPage.fillSelectReviewersForm([documentDetails.owner])
+      await documentContentPage.checkDocumentStatus(DocumentStatus.IN_REVIEW)
+      await documentContentPage.checkDocument({
+        ...documentDetails,
+        status: DocumentStatus.IN_REVIEW
+      })
+      await attachScreenshot('TESTS-139_send_for_review.png', page)
+    })
+
+    await test.step('4. Add comments and Complete Review', async () => {
+      await documentContentPage.addMessageToTheSectionTitle(newContentFirst.sectionTitle, messageToTitle)
+      await documentContentPage.addMessageToTheText(newContentFirst.content, messageToContent)
+      await documentContentPage.addMessageToTheSectionTitle(newContentSecond.sectionTitle, messageToSecondTitle)
+
+      await documentContentPage.buttonComments.click()
+
+      const documentCommentsPage = new DocumentCommentsPage(page)
+      await documentCommentsPage.checkCommentExist(newContentFirst.sectionTitle, 2)
+      await documentCommentsPage.checkCommentExist(newContentSecond.sectionTitle, 1)
+
+      await documentContentPage.completeReview()
+
+      await documentContentPage.checkDocumentStatus(DocumentStatus.REVIEWED)
+      await documentContentPage.checkCurrentRights(DocumentRights.VIEWING)
+      await attachScreenshot('TESTS-139_add_comments.png', page)
+    })
+
+    await test.step('5. Update Document and fix reviews', async () => {
+      await documentContentPage.buttonEditDocument.click()
+
+      await documentContentPage.updateSectionTitle('1', updateContentFirst.sectionTitle)
+      await documentContentPage.addContentToTheSection(updateContentFirst)
+      await documentContentPage.checkContentForTheSection(updateContentFirst)
+
+      const documentCommentsPage = new DocumentCommentsPage(page)
+      await documentCommentsPage.checkCommentExist(updateContentFirst.sectionTitle, 2)
+      await documentCommentsPage.checkCommentExist(newContentSecond.sectionTitle, 1)
+      await documentCommentsPage.resolveAllComments()
+
+      await documentCommentsPage.checkCommentNotExist(updateContentFirst.sectionTitle)
+      await documentCommentsPage.checkCommentNotExist(newContentSecond.sectionTitle)
+      await attachScreenshot('TESTS-136_fix_reviews.png', page)
+    })
+
+    await test.step('6. Check document status and details', async () => {
+      await documentContentPage.buttonDocumentInformation.click()
+      await documentContentPage.checkDocumentStatus(DocumentStatus.DRAFT)
+      await documentContentPage.checkDocument({
+        ...documentDetails,
+        status: DocumentStatus.DRAFT,
+        version: 'v0.1'
+      })
+      await attachScreenshot('TESTS-139_check_document.png', page)
+    })
+  })
+
+  test('TESTS-140. Comparing several document versions', async ({ page }) => {
+    await allure.description('Requirement\nUsers need to compare several document versions')
+    await allure.tms('TESTS-140', 'https://front.hc.engineering/workbench/platform/tracker/TESTS-140')
+    const makeReviewDocument: NewDocument = {
+      template: 'HR (HR)',
+      title: `Comparing document versions-${generateId()}`,
+      description: `Comparing document versions description-${generateId()}`
+    }
+    const documentDetails: DocumentDetails = {
+      type: 'HR',
+      category: 'Human Resources',
+      version: 'v0.1',
+      status: DocumentStatus.DRAFT,
+      owner: 'Appleseed John',
+      author: 'Appleseed John'
+    }
+    const newContentFirst: Content = {
+      sectionTitle: `Comparing versions. Overview-${generateId()}`,
+      content: `Comparing versions. New content-${generateId()}!!!!`
+    }
+    const updateContentFirst: Content = {
+      sectionTitle: '',
+      content: `Comparing versions Updated. Updated content-${generateId()}!!!!`
+    }
+    const newContentSecond: Content = {
+      sectionTitle: `Comparing versions. Description-${generateId()}`,
+      content: `Comparing versions. New content Description-${generateId()}!!!!`
+    }
+
+    await prepareDocumentStep(page, makeReviewDocument)
+    const documentContentPage = new DocumentContentPage(page)
+    await test.step('2. Add section and content', async () => {
+      await documentContentPage.checkDocumentTitle(makeReviewDocument.title)
+      await documentContentPage.checkDocumentStatus(DocumentStatus.DRAFT)
+      await documentContentPage.checkDocument(documentDetails)
+      await documentContentPage.checkCurrentRights(DocumentRights.EDITING)
+
+      await documentContentPage.updateSectionTitle('1', newContentFirst.sectionTitle)
+      await documentContentPage.addContentToTheSection(newContentFirst)
+      await attachScreenshot('TESTS-140_add_section_and_content.png', page)
+    })
+
+    await test.step('3. Send for Review', async () => {
+      await documentContentPage.buttonSendForReview.click()
+      await documentContentPage.fillSelectReviewersForm([documentDetails.owner])
+      await documentContentPage.checkDocumentStatus(DocumentStatus.IN_REVIEW)
+      await documentContentPage.checkDocument({
+        ...documentDetails,
+        status: DocumentStatus.IN_REVIEW
+      })
+      await attachScreenshot('TESTS-140_send_for_review.png', page)
+    })
+
+    await test.step('4. Add comments and Complete Review', async () => {
+      await documentContentPage.completeReview()
+      await documentContentPage.checkDocumentStatus(DocumentStatus.REVIEWED)
+      await documentContentPage.checkCurrentRights(DocumentRights.VIEWING)
+      await attachScreenshot('TESTS-140_add_comments.png', page)
+    })
+
+    await test.step('5. Compare the versions with several documents', async () => {
+      await documentContentPage.buttonEditDocument.click()
+      await documentContentPage.checkCurrentRights(DocumentRights.EDITING)
+      await documentContentPage.checkDocument({
+        ...documentDetails,
+        status: DocumentStatus.DRAFT,
+        version: 'v0.1'
+      })
+
+      await documentContentPage.addContentToTheSection({
+        sectionTitle: newContentFirst.sectionTitle,
+        content: updateContentFirst.content
+      })
+      await documentContentPage.checkContentForTheSection({
+        sectionTitle: newContentFirst.sectionTitle,
+        content: updateContentFirst.content
+      })
+
+      await documentContentPage.addNewSection('1', 'below')
+      await documentContentPage.updateSectionTitle('2', newContentSecond.sectionTitle)
+      await documentContentPage.addContentToTheSection(newContentSecond)
+
+      await documentContentPage.changeCurrentRight(DocumentRights.COMPARING)
+      await documentContentPage.checkComparingTextAdded('Updated')
+      await documentContentPage.checkComparingTextDeleted('New')
+      await documentContentPage.checkComparingTextAdded(newContentSecond.sectionTitle)
+      await documentContentPage.checkComparingTextAdded(newContentSecond.content)
+
+      await attachScreenshot('TESTS-140_fix_reviews.png', page)
+    })
+  })
+
+  test('TESTS-141. Send for approval document after resolve comments', async ({ page, browser }) => {
+    await allure.description(
+      'Requirement\nUsers need to make a resolve all comments and done documents for the Effective status'
+    )
+    await allure.tms('TESTS-141', 'https://front.hc.engineering/workbench/platform/tracker/TESTS-141')
+    const userSecondPage = await getSecondPage(browser)
+    const completeDocument: NewDocument = {
+      template: 'HR (HR)',
+      title: `Complete document-${generateId()}`,
+      description: `Complete document description-${generateId()}`
+    }
+    const reviewer = 'Dirak Kainin'
+    const documentDetails: DocumentDetails = {
+      type: 'HR',
+      category: 'Human Resources',
+      version: 'v0.1',
+      status: DocumentStatus.DRAFT,
+      owner: 'Appleseed John',
+      author: 'Appleseed John'
+    }
+    const newContentFirst: Content = {
+      sectionTitle: `Complete document. Overview-${generateId()}`,
+      content: `Complete document. New content-${generateId()}!!!!`
+    }
+    const updateContentFirst: Content = {
+      sectionTitle: `Complete document Updated. Updated Overview-${generateId()}`,
+      content: `Complete document Updated. Updated content-${generateId()}!!!!`
+    }
+    const newContentSecond: Content = {
+      sectionTitle: `Complete document. Description-${generateId()}`,
+      content: `Complete document. New content Description-${generateId()}!!!!`
+    }
+    const messageToTitle: string = `Complete document. Message to the first title-${generateId()}`
+    const messageToContent: string = `Complete document. Message to the content-${generateId()}`
+
+    await prepareDocumentStep(page, completeDocument)
+
+    const documentContentPage = new DocumentContentPage(page)
+    await test.step('2. Add section and content', async () => {
+      await documentContentPage.checkDocumentTitle(completeDocument.title)
+      await documentContentPage.checkDocumentStatus(DocumentStatus.DRAFT)
+      await documentContentPage.checkDocument(documentDetails)
+      await documentContentPage.checkCurrentRights(DocumentRights.EDITING)
+
+      await documentContentPage.updateSectionTitle('1', newContentFirst.sectionTitle)
+      await documentContentPage.addContentToTheSection(newContentFirst)
+
+      await documentContentPage.addNewSection('1', 'below')
+      await documentContentPage.updateSectionTitle('2', newContentSecond.sectionTitle)
+      await documentContentPage.addContentToTheSection(newContentSecond)
+      await attachScreenshot('TESTS-141_add_section_and_content.png', page)
+    })
+
+    await test.step('3. Send for Review', async () => {
+      await documentContentPage.buttonSendForReview.click()
+      await documentContentPage.fillSelectReviewersForm([reviewer])
+      await documentContentPage.checkDocumentStatus(DocumentStatus.IN_REVIEW)
+      await documentContentPage.checkDocument({
+        ...documentDetails,
+        status: DocumentStatus.IN_REVIEW
+      })
+      await attachScreenshot('TESTS-141_send_for_review.png', page)
+    })
+
+    await test.step('4. As author add comments to the first section', async () => {
+      await documentContentPage.addMessageToTheSectionTitle(newContentFirst.sectionTitle, messageToTitle)
+      await documentContentPage.addMessageToTheText(newContentFirst.content, messageToContent)
+
+      await documentContentPage.buttonComments.click()
+
+      const documentCommentsPage = new DocumentCommentsPage(page)
+      await documentCommentsPage.checkCommentExist(newContentFirst.sectionTitle, 2)
+      await documentCommentsPage.checkCommentCanBeResolved(newContentFirst.sectionTitle, 1)
+      await documentCommentsPage.checkCommentCanBeResolved(newContentFirst.sectionTitle, 2)
+      await attachScreenshot('TESTS-141_author_add_comments.png', page)
+    })
+
+    await test.step('5. As reviewer add comments to the second section and Complete Review', async () => {
+      await (await userSecondPage.goto(`${PlatformURI}/${DocumentURI}`))?.finished()
+
+      const documentsPageSecond = new DocumentsPage(userSecondPage)
+      await documentsPageSecond.openDocument(completeDocument.title)
+
+      const documentContentPageSecond = new DocumentContentPage(userSecondPage)
+      await documentContentPageSecond.addMessageToTheSectionTitle(newContentSecond.sectionTitle, messageToTitle)
+      await documentContentPageSecond.addMessageToTheText(newContentSecond.content, messageToContent)
+
+      await documentContentPageSecond.buttonComments.click()
+
+      const documentCommentsPageSecond = new DocumentCommentsPage(userSecondPage)
+      await documentCommentsPageSecond.checkCommentExist(newContentSecond.sectionTitle, 2)
+      await documentCommentsPageSecond.checkCommentCanBeResolved(newContentSecond.sectionTitle, 3)
+      await documentCommentsPageSecond.checkCommentCanBeResolved(newContentSecond.sectionTitle, 4)
+
+      // // TODO uncomment after fix https://front.hc.engineering/workbench/platform/tracker/EZQMS-552
+      // await documentCommentsPageSecond.checkCommentCanNotBeResolved(newContentFirst.sectionTitle, 1)
+      // await documentCommentsPageSecond.checkCommentCanNotBeResolved(newContentFirst.sectionTitle, 2)
+
+      await documentContentPageSecond.completeReview()
+
+      await documentContentPageSecond.checkDocumentStatus(DocumentStatus.REVIEWED)
+      await documentContentPageSecond.checkCurrentRights(DocumentRights.VIEWING)
+      await attachScreenshot('TESTS-141_reviewer_add_comments.png', page)
+    })
+
+    await test.step('6. Update Document and fix reviews', async () => {
+      await documentContentPage.buttonEditDocument.click()
+
+      await documentContentPage.updateSectionTitle('1', updateContentFirst.sectionTitle)
+      await documentContentPage.addContentToTheSection(updateContentFirst)
+      await documentContentPage.checkContentForTheSection(updateContentFirst)
+
+      const documentCommentsPage = new DocumentCommentsPage(page)
+      await documentCommentsPage.checkCommentExist(updateContentFirst.sectionTitle, 2)
+      await documentCommentsPage.resolveAllComments()
+
+      await documentCommentsPage.checkCommentNotExist(updateContentFirst.sectionTitle)
+      await attachScreenshot('TESTS-141_fix_reviews.png', page)
+
+      await documentContentPage.buttonDocumentInformation.click()
+      await documentContentPage.checkDocumentStatus(DocumentStatus.DRAFT)
+      await documentContentPage.checkDocument({
+        ...documentDetails,
+        status: DocumentStatus.DRAFT,
+        version: 'v0.1'
+      })
+      await attachScreenshot('TESTS-141_check_document.png', page)
+    })
+
+    await test.step('7. Send for Approval', async () => {
       await documentContentPage.buttonSendForApproval.click()
-      await documentContentPage.fillSelectApproversForm([documentDetails.owner])
+      await documentContentPage.fillSelectApproversForm([reviewer])
       await documentContentPage.checkDocumentStatus(DocumentStatus.IN_APPROVAL)
       await documentContentPage.checkDocument({
         ...documentDetails,
-        status: DocumentStatus.IN_APPROVAL
+        status: DocumentStatus.IN_APPROVAL,
+        version: 'v0.1'
       })
       await documentContentPage.checkCurrentRights(DocumentRights.VIEWING)
     })
 
-    await test.step('3. Approve document', async () => {
-      await documentContentPage.confirmApproval()
+    await test.step('8. Approve document', async () => {
+      const documentsPageSecond = new DocumentsPage(userSecondPage)
+      await documentsPageSecond.openDocument(completeDocument.title)
+
+      const documentContentPageSecond = new DocumentContentPage(userSecondPage)
+      await documentContentPageSecond.confirmApproval()
+
+      await documentContentPageSecond.buttonDocumentInformation.click()
+      await documentContentPageSecond.checkDocumentStatus(DocumentStatus.EFFECTIVE)
+      await documentContentPageSecond.checkDocument({
+        ...documentDetails,
+        status: DocumentStatus.EFFECTIVE,
+        version: 'v0.1'
+      })
+      await documentContentPageSecond.checkCurrentRights(DocumentRights.VIEWING)
+
+      await attachScreenshot('TESTS-141_approve_document.png', page)
     })
 
-    await test.step('4. Check the document and status', async () => {
+    await test.step('9. Check document', async () => {
       await documentContentPage.checkDocumentStatus(DocumentStatus.EFFECTIVE)
       await documentContentPage.checkDocument({
         ...documentDetails,
@@ -422,10 +800,15 @@ test.describe('QMS. Documents tests', () => {
       })
       await documentContentPage.checkCurrentRights(DocumentRights.VIEWING)
 
-      await documentContentPage.openApprovals()
-      const documentApprovalsPage = new DocumentApprovalsPage(page)
-      await documentApprovalsPage.checkSuccessApproval(documentDetails.owner)
-      await attachScreenshot('TESTS-137_approve_document.png', page)
+      await attachScreenshot('TESTS-141_check_document.png', page)
+    })
+
+    await test.step('10. Check History tab', async () => {
+      await documentContentPage.buttonHistoryTab.click()
+
+      const documentHistoryPage = new DocumentHistoryPage(page)
+      await documentHistoryPage.checkHistoryEventExist('New document creation')
+      await attachScreenshot('TESTS-141_check_history_tab.png', page)
     })
   })
