+++ qms-tests/sanity/tsconfig.json
@@ -1,8 +1,9 @@
 {
   "extends": "./node_modules/@hcengineering/platform-rig/profiles/default/tsconfig.json",
   "compilerOptions": {
-    "rootDir": "./qms-tests",
+    "rootDir": "./tests",
     "outDir": "./lib",
+    "declarationDir": "./types",
     "lib": ["esnext", "dom"]
   }
 }
